filestructure refactor
