+++ src/logger/LogEntry.h
